# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clematis Storage API",
    "description": "Clematis Storage API",
    "version": "1.0.0",
    "summary": "Hateoas Restful API for a file storage"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/files/upload/batch": {
      "post": {
        "tags": [
          "file-storage-controller"
        ],
        "summary": "Upload multiple files",
        "description": "Uploads multiple files and optionally stores them under the specified path.",
        "operationId": "uploadBatch",
        "parameters": [
          {
            "name": "files",
            "in": "query",
            "description": "List of files to upload",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Optional storage path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files successfully uploaded",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/db/upload/batch": {
      "post": {
        "tags": [
          "storage-controller"
        ],
        "summary": "Upload multiple files",
        "description": "Uploads multiple files and optionally stores them under the specified path.",
        "operationId": "uploadBatch_1",
        "parameters": [
          {
            "name": "files",
            "in": "query",
            "description": "List of files to upload",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Optional storage path",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files successfully uploaded",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/metadata/{id}": {
      "get": {
        "tags": [
          "file-storage-controller"
        ],
        "summary": "Get metadata only",
        "description": "Returns only metadata for the stored file (no binary content).",
        "operationId": "getMetadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata returned",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/getByPath": {
      "get": {
        "tags": [
          "file-storage-controller"
        ],
        "summary": "Get files by path prefix",
        "description": "Returns all files whose stored names start with the given prefix. If prefix is empty or null, returns all files.",
        "operationId": "getByPath",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of files returned successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/find": {
      "get": {
        "tags": [
          "file-storage-controller"
        ],
        "summary": "Find files by path prefix",
        "description": "Returns stored file metadata whose logical path starts with the provided prefix. If path is blank or missing, returns all files.",
        "operationId": "findByPath",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Path prefix to search",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matched files",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/db/metadata/{id}": {
      "get": {
        "tags": [
          "storage-controller"
        ],
        "summary": "Get metadata only",
        "description": "Returns only metadata for the stored file (no binary content).",
        "operationId": "getMetadata_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata returned",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/db/getByPath": {
      "get": {
        "tags": [
          "storage-controller"
        ],
        "summary": "Get files by path prefix",
        "description": "Returns all files whose stored names start with the given prefix. If prefix is empty or null, returns all files.",
        "operationId": "getByPath_1",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of files returned successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/db/find": {
      "get": {
        "tags": [
          "storage-controller"
        ],
        "summary": "Find files by path prefix",
        "description": "Returns stored file metadata whose logical path starts with the provided prefix. If path is blank or missing, returns all files.",
        "operationId": "findByPath_1",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Path prefix to search",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matched files",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/{id}": {
      "delete": {
        "tags": [
          "file-storage-controller"
        ],
        "summary": "Delete file by ID",
        "description": "Deletes a file from the storage by its ID.",
        "operationId": "deleteFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the file to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/db/{id}": {
      "delete": {
        "tags": [
          "storage-controller"
        ],
        "summary": "Delete file by ID",
        "description": "Deletes a file from the storage by its ID.",
        "operationId": "deleteFile_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the file to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RequestResponse": {
        "type": "object",
        "description": "Response returned after a successful upload",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Original filename (or logical stored filename)",
            "example": "photo.png"
          },
          "downloadUrl": {
            "type": "string",
            "description": "Public URL that can be used to download the file",
            "example": "http://localhost:8080/api/files/1234"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "image/png"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "File size in bytes",
            "example": 34567
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error wrapper",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human readable message",
            "example": "File not found"
          },
          "status": {
            "type": "string",
            "description": "HTTP status name",
            "example": "NOT_FOUND"
          }
        }
      },
...
    }
  }
}